Attempt to deal with many complex Node.js and node-sass installation issues.

+++ build.sh
@@ -15,9 +15,16 @@ if [ "$EUID" != 0 ]; then
 fi
 
 sudo -u "$SUDO_USER" bash -c ./build_node_check.sh
-path="$(cat node_path.txt)"
-rm node_path.txt
+
+if [ -f "node_path.txt" ]; then
+  path="--path \"$(cat node_path.txt)\""
+  rm node_path.txt
+
+  if [ "$path" == "failed" ]; then
+    exit;
+  fi
+fi
 
 echo "Starting main installer..."
-# shellcheck disable=SC2048,SC2086 # this should become separate items with spaces in between, not one quoted thing
-npm run build:prod -- --path "$path" --bash $*
+# shellcheck disable=SC2048,SC2086 #
+npm run build:prod "$path" --bash $*
